requirements.txt
streamlit
Pillow
io
requests